move default_rotation
