+++ apps/src/sketchlab/reactFlow/elementToolbars/TextNodeToolbar.tsx
@@ -5,7 +5,9 @@ import {TextNodeType} from '../types';
 import ActionsGroup from './ActionsGroup';
 import FontSizeGroup from './FontSizeGroup';
 import HandleVisibilityToggle from './HandleVisibilityToggle';
-import RotationGroup, {DEFAULT_ROTATION} from './RotationGroup';
+import {DEFAULT_ROTATION} from '../constants';
+
+import RotationGroup from './RotationGroup';
 import SwatchGroup from './SwatchGroup';
 import {
   DEFAULT_FONT_COLOR,
